+++ src/Concordium/Client/Commands.hs
@@ -13,6 +13,9 @@ module Concordium.Client.Commands (
     Interval (..),
     InteractionOpts (..),
     TransactionCmd (..),
+    PLTCmd (..),
+    TokenSupplyAction (..),
+    ModifyListAction (..),
     AccountCmd (..),
     ModuleCmd (..),
     ContractCmd (..),
@@ -223,12 +226,34 @@ data TransactionCmd
           -- | Options for transaction.
           trdTransactionOptions :: !(TransactionOpts (Maybe Energy))
         }
-    | TransactionPLTTransfer
-        { tthReceiver :: !Text,
-          tthAmount :: !TokenAmount,
-          tthSymbol :: !Text,
-          tthMemo :: !(Maybe MemoInput),
-          tthOpts :: !(TransactionOpts (Maybe Energy))
+    | PLTCmd PLTCmd
+    deriving (Show)
+
+data TokenSupplyAction = Mint | Burn
+    deriving (Show, Eq)
+
+data ModifyListAction = AddAllowList | RemoveAllowList | AddDenyList | RemoveDenyList
+    deriving (Show, Eq)
+
+data PLTCmd
+    = TransactionPLTTransfer
+        { tptReceiver :: !Text,
+          tptAmount :: !TokenAmount,
+          tptSymbol :: !Text,
+          tptMemo :: !(Maybe MemoInput),
+          tptOpts :: !(TransactionOpts (Maybe Energy))
+        }
+    | TransactionPLTUpdateSupply
+        { tpusAction :: !TokenSupplyAction,
+          tpusAmount :: !TokenAmount,
+          tpusSymbol :: !Text,
+          tpusOpts :: !(TransactionOpts (Maybe Energy))
+        }
+    | TransactionPLTModifyList
+        { tpmlAction :: !ModifyListAction,
+          tpmlAccount :: !Text,
+          tpmlSymbol :: !Text,
+          tpmlOpts :: !(TransactionOpts (Maybe Energy))
         }
     deriving (Show)
 
@@ -720,10 +745,29 @@ transactionCmds =
                         <> transactionWithScheduleCmd
                         <> transactionDeployCredentialCmd
                         <> transactionRegisterDataCmd
-                        <> transactionPLTTransferCmd
+                        <> pltCmds
                     )
             )
             (progDesc "Commands for submitting and inspecting transactions.")
+        )
+
+pltCmds :: Mod CommandFields TransactionCmd
+pltCmds =
+    command
+        "plt"
+        ( info
+            ( PLTCmd
+                <$> hsubparser
+                    ( transactionPLTTransferCmd
+                        <> transactionPLTMintCmd
+                        <> transactionPLTBurnCmd
+                        <> transactionPLTAddAllowListCmd
+                        <> transactionPLTAddDenyListCmd
+                        <> transactionPLTRemoveAllowListCmd
+                        <> transactionPLTRemoveDenyListCmd
+                    )
+            )
+            (progDesc "Commands for PLTs (protocol level tokens) transactions.")
         )
 
 transactionSubmitCmd :: Mod CommandFields TransactionCmd
@@ -835,19 +879,97 @@ transactionSendCcdCmd =
             (progDesc "Transfer CCD from one account to another.")
         )
 
-transactionPLTTransferCmd :: Mod CommandFields TransactionCmd
+transactionPLTTransferCmd :: Mod CommandFields PLTCmd
 transactionPLTTransferCmd =
     command
-        "transfer-plt"
+        "send"
         ( info
             ( TransactionPLTTransfer
                 <$> strOption (long "receiver" <> metavar "RECEIVER-ACCOUNT" <> help "Address of the receiver.")
                 <*> option (eitherReader tokenAmountFromStringInform) (long "amount" <> metavar "TOKEN-AMOUNT" <> help "Amount of tokens to send.")
-                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (Symbol) of the token.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
                 <*> memoInputParser
                 <*> transactionOptsParser
             )
             (progDesc "Transfer tokens from one account to another.")
+        )
+
+transactionPLTMintCmd :: Mod CommandFields PLTCmd
+transactionPLTMintCmd =
+    command
+        "mint"
+        ( info
+            ( TransactionPLTUpdateSupply Mint
+                <$> option (eitherReader tokenAmountFromStringInform) (long "amount" <> metavar "TOKEN-AMOUNT" <> help "Amount of tokens to send.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Mint PLTs (protocol level tokens).")
+        )
+
+transactionPLTBurnCmd :: Mod CommandFields PLTCmd
+transactionPLTBurnCmd =
+    command
+        "burn"
+        ( info
+            ( TransactionPLTUpdateSupply Burn
+                <$> option (eitherReader tokenAmountFromStringInform) (long "amount" <> metavar "TOKEN-AMOUNT" <> help "Amount of tokens to send.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Burn PLTs (protocol level tokens).")
+        )
+
+transactionPLTAddAllowListCmd :: Mod CommandFields PLTCmd
+transactionPLTAddAllowListCmd =
+    command
+        "add-to-allow-list"
+        ( info
+            ( TransactionPLTModifyList AddAllowList
+                <$> strOption (long "account" <> metavar "ACCOUNT" <> help "The account to add to the list.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Add an account to the allow list.")
+        )
+
+transactionPLTAddDenyListCmd :: Mod CommandFields PLTCmd
+transactionPLTAddDenyListCmd =
+    command
+        "add-to-deny-list"
+        ( info
+            ( TransactionPLTModifyList AddDenyList
+                <$> strOption (long "account" <> metavar "ACCOUNT" <> help "The account to add to the list.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Add an account to the deny list.")
+        )
+
+transactionPLTRemoveAllowListCmd :: Mod CommandFields PLTCmd
+transactionPLTRemoveAllowListCmd =
+    command
+        "remove-from-allow-list"
+        ( info
+            ( TransactionPLTModifyList RemoveAllowList
+                <$> strOption (long "account" <> metavar "ACCOUNT" <> help "The account to remove from the list.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Remove an account from the allow list.")
+        )
+
+transactionPLTRemoveDenyListCmd :: Mod CommandFields PLTCmd
+transactionPLTRemoveDenyListCmd =
+    command
+        "remove-from-deny-list"
+        ( info
+            ( TransactionPLTModifyList RemoveDenyList
+                <$> strOption (long "account" <> metavar "ACCOUNT" <> help "The account to remove from the list.")
+                <*> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (symbol) of the token.")
+                <*> transactionOptsParser
+            )
+            (progDesc "Remove an account from the deny list.")
         )
 
 transactionWithScheduleCmd :: Mod CommandFields TransactionCmd