+++ ChangeLog.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 9.1.0 - 2025-05-20
+
+- Add `transaction plt add-to-allow-list` to add an account to the allow list by the governance account.
+- Add `transaction plt add-to-deny-list` to add an account to the deny list by the governance account.
+- Add `transaction plt remove-from-allow-list` to remove an account from the allow list by the governance account.
+- Add `transaction plt remove-from-deny-list` to remove an account from the deny list by the governance account.
+- Add `transaction plt mint` command to mint protocol level tokens by the governance account.
+- Add `transaction plt burn` command to burn protocol level tokens by the governance account.
+- Refactore plt commands into its own subcommand.
+
 ## 9.0.0 - 2025-05-09
 
 - Add `raw GetTokenInfo` command to display the info of a protocol level token.